auth: merge auth_tenant_repo into auth_repo
